# swagger.yaml
definitions:
  CreateAwsIamUserSessionRequest:
    properties:
      accountNumber:
        description: the account number of the aws account related to the role
        type: string
        x-go-name: AccountNumber
      awsAccessKeyId:
        type: string
        x-go-name: AwsAccessKeyId
      awsSecretAccessKey:
        type: string
        x-go-name: AwsSecretAccessKey
      mfaDevice:
        type: string
        x-go-name: MfaDevice
      name:
        description: the name which will be displayed
        type: string
        x-go-name: Name
      profileName:
        type: string
        x-go-name: ProfileName
      region:
        description: the region on which the session will be initialized
        type: string
        x-go-name: Region
      user:
        type: string
        x-go-name: User
    required:
    - name
    - accountNumber
    - region
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/request_dto/plain_aws_session_dto
  CreateAwsIamRoleChainedSessionRequestDto:
    properties:
      accountName:
        description: the name which will be displayed
        type: string
        x-go-name: AccountName
      accountNumber:
        description: the account number of the aws account related to the role
        type: string
        x-go-name: AccountNumber
      parentId:
        description: |-
          the parent session id, can be an aws plain or federated session
          it's generated with an uuid v4
        type: string
        x-go-name: ParentId
      region:
        description: the region on which the session will be initialized
        type: string
        x-go-name: Region
      roleName:
        description: the role name
        type: string
        x-go-name: RoleName
    required:
    - parentId
    - accountName
    - accountNumber
    - roleName
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/request_dto/trusted_aws_session_dto
  EditAwsIamRoleChainedSessionRequestDto:
    properties:
      accountName:
        description: the name which will be displayed
        type: string
        x-go-name: AccountName
      accountNumber:
        description: the account number of the aws account related to the role
        type: string
        x-go-name: AccountNumber
      parentId:
        description: |-
          the parent session id, can be an aws plain or federated session
          it's generated with an uuid v4
        type: string
        x-go-name: ParentId
      region:
        description: the region on which the session will be initialized
        type: string
        x-go-name: Region
      roleName:
        description: the role name
        type: string
        x-go-name: RoleName
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/request_dto/trusted_aws_session_dto
  GetAwsIamUserSessionResponse:
    properties:
      Data:
        $ref: '#/definitions/AwsIamUserSession'
      Message:
        type: string
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/response_dto/plain_aws_session_dto
  MessageResponse:
    properties:
      Message:
        type: string
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/response_dto
  PlainAwsAccount:
    properties:
      MfaDevice:
        type: string
      NamedProfileId:
        type: string
      Region:
        type: string
      SessionTokenExpiration:
        type: string
    type: object
    x-go-package: leapp_daemon/domain/session
  AwsIamUserSession:
    properties:
      Account:
        $ref: '#/definitions/PlainAwsAccount'
      Alias:
        type: string
      Id:
        type: string
      LastStopTime:
        type: string
      StartTime:
        type: string
      Status:
        $ref: '#/definitions/Status'
    type: object
    x-go-package: leapp_daemon/domain/session
  Status:
    format: int64
    type: integer
    x-go-package: leapp_daemon/domain/session
  TrustedAwsAccount:
    properties:
      AccountNumber:
        type: string
      Name:
        type: string
      Region:
        type: string
      Role:
        $ref: '#/definitions/TrustedAwsRole'
    type: object
    x-go-package: leapp_daemon/domain/session
  TrustedAwsRole:
    properties:
      Arn:
        type: string
      Name:
        type: string
    type: object
    x-go-package: leapp_daemon/domain/session
  AwsIamRoleChainedSession:
    properties:
      Account:
        $ref: '#/definitions/TrustedAwsAccount'
      Id:
        type: string
      ParentId:
        type: string
      StartTime:
        type: string
      Status:
        $ref: '#/definitions/Status'
    type: object
    x-go-package: leapp_daemon/domain/session
  UpdateAwsIamUserSessionRequest:
    properties:
      accountNumber:
        type: string
        x-go-name: AccountNumber
      awsAccessKeyId:
        type: string
        x-go-name: AwsAccessKeyId
      awsSecretAccessKey:
        type: string
        x-go-name: AwsSecretAccessKey
      mfaDevice:
        type: string
        x-go-name: MfaDevice
      name:
        type: string
        x-go-name: Name
      profileName:
        type: string
        x-go-name: ProfileName
      region:
        type: string
        x-go-name: Region
      user:
        type: string
        x-go-name: User
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/request_dto/plain_aws_session_dto
  UpdateAwsIamUserSessionUriRequest:
    properties:
      Id:
        type: string
    type: object
    x-go-package: leapp_daemon/interface/http/controller/dto/request_dto/plain_aws_session_dto
  getAwsIamRoleChainedSessionResponse:
    properties:
      Data:
        $ref: '#/definitions/AwsIamRoleChainedSession'
      Message:
        type: string
    type: object
    x-go-package: leapp_daemon/interface/http/controller
paths:
  /session/plain:
    post:
      description: Create a new AWS Iam User Session
      operationId: createAwsIamUserSession
      parameters:
      - description: AWS Iam User session create body
        in: body
        name: Body
        schema:
          $ref: '#/definitions/CreateAwsIamUserSessionRequest'
      responses:
        "200":
          description: MessageResponse
          schema:
            $ref: '#/definitions/MessageResponse'
      tags:
      - awsIamUserSession
  /session/plain/{id}:
    delete:
      description: Delete a AWS Iam User Session
      operationId: deleteAwsIamUserSession
      responses:
        "200":
          description: MessageResponse
          schema:
            $ref: '#/definitions/MessageResponse'
      tags:
      - awsIamUserSession
    get:
      description: Get a AWS Iam User Session
      operationId: getAwsIamUserSession
      responses:
        "200":
          description: GetAwsIamUserSessionResponse
          schema:
            $ref: '#/definitions/GetAwsIamUserSessionResponse'
      tags:
      - awsIamUserSession
    put:
      description: Edit a AWS Iam User Session
      operationId: updateAwsIamUserSession
      parameters:
      - description: AWS Iam User session update uri body
        in: body
        name: Body
        schema:
          $ref: '#/definitions/UpdateAwsIamUserSessionRequest'
      responses:
        "200":
          description: MessageResponse
          schema:
            $ref: '#/definitions/MessageResponse'
      tags:
      - awsIamUserSession
  /session/plain/{id}/start:
    post:
      description: Start a AWS Iam User Session
      operationId: startAwsIamUserSession
      responses:
        "200":
          description: MessageResponse
          schema:
            $ref: '#/definitions/MessageResponse'
      tags:
      - awsIamUserSession
  /session/trusted:
    post:
      description: Create a new trusted aws session
      operationId: createAwsIamRoleChainedSession
      parameters:
      - description: This text will appear as description of your request body.
        in: body
        name: Body
        schema:
          $ref: '#/definitions/CreateAwsIamRoleChainedSessionRequestDto'
      responses:
        "200":
          $ref: '#/responses/messageResponse'
      tags:
      - session-trusted-aws
  /session/trusted/{id}:
    delete:
      description: Delete a Trusted AWS Session
      operationId: deleteAwsIamRoleChainedSession
      parameters:
      - description: the id of the trusted aws session
        in: path
        name: Id
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/messageResponse'
      tags:
      - session-trusted-aws
    get:
      description: Get a Trusted AWS Session
      operationId: getAwsIamRoleChainedSession
      parameters:
      - description: the id of the trusted aws session
        in: path
        name: Id
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/getAwsIamRoleChainedSessionResponse'
      tags:
      - session-trusted-aws
    put:
      description: Edit a Trusted AWS Session
      operationId: editAwsIamRoleChainedSession
      parameters:
      - description: This text will appear as description of your request body.
        in: body
        name: Body
        schema:
          $ref: '#/definitions/EditAwsIamRoleChainedSessionRequestDto'
      - description: the id of the trusted aws session
        in: path
        name: Id
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/messageResponse'
      tags:
      - session-trusted-aws
responses:
  getAwsIamUserSessionResponse:
    description: ""
    schema:
      $ref: '#/definitions/GetAwsIamUserSessionResponse'
  getAwsIamRoleChainedSessionResponse:
    description: ""
    schema:
      $ref: '#/definitions/getAwsIamRoleChainedSessionResponse'
  messageResponse:
    description: ""
    schema:
      $ref: '#/definitions/MessageResponse'
swagger: "2.0"